backend-api-server/requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 46.0.0
sqlalchemy == 1.3.16
SQLAlchemy-Enum-List == 0.1.1
checksumdir == 1.1.7
numpy == 1.16.4
url-normalize ==1.4.2
xlrd==1.2.0
psycopg2-binary==2.8.4
w3lib==1.21.0
pandas==1.1.2
requests==2.22.0
flask==1.1.2
flask-cors == 3.0.9
pyopenssl
psycopg2-binary==2.8.4
pycryptodome==3.9.8
mal2-fakeshop-models == 0.25